WebSockets/tests/submissions/Opera/interfaces/WebSocket/events/020.html

<!doctype html>
<title>WebSockets: error events</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js></script>
<div id=log></div>
<script>
async_test(function(t) {
  var ws = new WebSocket(SCHEME_AND_DOMAIN+':9000/echo');
  ws.onerror = t.step_func(function(e) {
    assert_true(e instanceof Event);
    t.done();
  })
});
</script>
